+++ src/util/virthreadwin32.c
@@ -33,7 +33,7 @@ typedef struct virThreadLocalData virThreadLocalData;
 typedef virThreadLocalData *virThreadLocalDataPtr;
 
 virMutex virThreadLocalLock;
-unsigned int virThreadLocalCount = 0;
+size_t virThreadLocalCount = 0;
 virThreadLocalDataPtr virThreadLocalList = NULL;
 DWORD selfkey;
 
@@ -54,7 +54,7 @@ int virThreadInitialize(void)
 
 void virThreadOnExit(void)
 {
-    unsigned int i;
+    size_t i;
     virMutexLock(&virThreadLocalLock);
     for (i = 0; i < virThreadLocalCount; i++) {
         if (virThreadLocalList[i].cleanup) {
@@ -219,7 +219,7 @@ void virCondBroadcast(virCondPtr c)
     virMutexLock(&c->lock);
 
     if (c->nwaiters) {
-        unsigned int i;
+        size_t i;
         for (i = 0; i < c->nwaiters; i++) {
             HANDLE event = c->waiters[i];
             SetEvent(event);
